+++ src/losant/contains-any.js
@@ -1,4 +1,6 @@
-const { curry, pipe, intersection } = require('ramda');
+const curry = require('ramda/src/curry');
+const pipe = require('ramda/src/pipe');
+const intersection = require('ramda/src/intersection');
 
 const isNotEmpty = require('./is-not-empty');
 
@@ -6,7 +8,7 @@ const isNotEmpty = require('./is-not-empty');
  * Takes two arrays and returns true if any of the values in the first array
  * are found in the second array.
  *
- * @signature Array -> Array -> Boolean
+ * @signature Array<* a> -> Array<* a> -> Boolean
  */
 const containsAny = curry((needleList, haystackList) =>
   pipe(