+++ src/losant/evolve-array.js
@@ -1,12 +1,11 @@
-const {
-  zipWith, call,
-} = require('ramda');
-
-// TODO: TESTS
+const zipWith = require('ramda/src/zipWith');
+const call = require('ramda/src/call');
 
 /**
  * Applies an array of transform functions to an array of values, with each
  * transform being applied to the value that has the same index.
+ *
+ * TODO: TESTS
  *
  * @signature [a -> b] -> [a] -> [b]
  *