+++ src/losant/is-nil-or-empty.js
@@ -1,9 +1,12 @@
-const {
-  either, isNil, isEmpty,
-} = require('ramda');
+const either = require('ramda/src/either');
+const isNil = require('ramda/src/isNil');
+const isEmpty = require('ramda/src/isEmpty');
 
 /**
- * Checks if a value is null, undefined, or an empty string.
+ * Checks if a value is null, undefined, an empty string, an empty object, an empty array.
+ *
+ * @see R.isNil
+ * @see R.isEmpty
  *
  * @signature * -> Boolean
  */