+++ src/losant/prop-or-strict.js
@@ -1,16 +1,20 @@
-const {
-  curry, compose, prop,
-} = require('ramda');
+const curry = require('ramda/src/curry');
+const compose = require('ramda/src/compose');
+const prop = require('ramda/src/prop');
+
 const defaultToStrict = require('./default-to-strict');
 
-// TODO: TESTS
-// TODO: DOCS
-
-// propOrStrict :: * a => String => Object -> a|*
-const propOrStrict = curry((def, key, data) => {
-  return compose(
+/**
+ * TODO: TESTS
+ * TODO: DOCS
+ *
+ * @signature * a -> String -> Object -> a|*
+ */
+const propOrStrict = curry((def, key, data) =>
+  compose(
     defaultToStrict(def),
     prop(key),
-  )(data);
-});
+  )(data)
+);
+
 module.exports = propOrStrict;