+++ src/losant/within.js
@@ -1,6 +1,12 @@
-const { curry, gt, lt, __, both } = require('ramda');
+const __ = require('ramda/src/__');
+const curry = require('ramda/src/curry');
+const gt = require('ramda/src/gt');
+const lt = require('ramda/src/lt');
+const both = require('ramda/src/both');
 
-// within :: Number -> Number -> Number -> Boolean
+/**
+ * @signature Number -> Number -> Number -> Boolean
+ */
 const within = curry(
   (start, end, value) =>
     both(