+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 Below should serve as an "upgrade guide" jumping off point as you are migrating projects to newer versions of this package. Ideally, interface changes will be called out specifically and one should not have to go sifting through git diffs to discern what changed.
 
 - `?.?.?` - *WIP*
+   - All internal losant utilities now only require the specific funtion they need from `ramda`/`lodash`.
    - Removed methods:
       - `indexBy` from Ramda has been omitted in favor of `keyBy`
+   - Added methods:
+      - `toPlainObject`
 - `1.1.2`
    - Added methods:
       - `keyByWith`